svsm_request: Move {clr,set}_vmsa_efer_svme() to cpu/vmsa.rs
No functional change intended.
